adding messages, views add policy from insured and add policy from coverage

+++ insurance_project/insurance/views.py
@@ -3,7 +3,7 @@ from django.contrib.auth import login
 from django.contrib import messages
 from django.shortcuts import render, get_object_or_404, redirect
 from .models import InsuredPerson, InsuranceType, InsuranceCoverage, Policy
-from .forms import InsuredPersonForm, InsuranceTypeForm, InsuranceCoverageForm, PolicyForm
+from .forms import InsuredPersonForm, InsuranceTypeForm, InsuranceCoverageForm, PolicyFormFromInsured, PolicyFormFromCoverage
 from django.core.paginator import Paginator
 
 
@@ -30,6 +30,7 @@ def insured_detail(request, id):
         'total_premium': total_premium,
     })
 
+
 # Přidání nového pojištěnce
 def add_insured(request):
     if request.method == 'POST':
@@ -50,6 +51,7 @@ def edit_insured(request, id):
         form = InsuredPersonForm(request.POST, instance=insured_person)
         if form.is_valid():
             form.save()
+            messages.success(request, 'Informace o pojištěnci byly úspěšně aktualizovány.')
             return redirect('insured_detail', id=id)
     else:
         form = InsuredPersonForm(instance=insured_person)
@@ -59,8 +61,11 @@ def edit_insured(request, id):
 # Odstranění pojištěnce
 def delete_insured(request, id):
     insured_person = get_object_or_404(InsuredPerson, id=id)
-    insured_person.delete()
-    return redirect('index')
+    if request.method == 'POST':
+        insured_person.delete()
+        messages.success(request, 'Pojištěnec byl úspěšně odebrán.')
+        return redirect('insured_list')
+    return render(request, 'insurance/confirm_delete.html', {'object': insured_person})
 
 
 # Seznam pojištění
@@ -81,6 +86,7 @@ def add_insurance(request):
         form = InsuranceTypeForm(request.POST)
         if form.is_valid():
             form.save()
+            messages.success(request, 'Nový typ pojištění byl úspěšně přidán.')
             return redirect('insurance_list')
     else:
         form = InsuranceTypeForm()
@@ -94,6 +100,7 @@ def edit_insurance(request, id):
         form = InsuranceTypeForm(request.POST, instance=insurance)
         if form.is_valid():
             form.save()
+            messages.success(request, 'Typ pojištění byl úspěšně aktualizován.')
             return redirect('insurance_list')
     else:
         form = InsuranceTypeForm(instance=insurance)
@@ -105,7 +112,7 @@ def delete_insurance(request, id):
 
     if request.method == 'POST':
         insurance_type.delete()
-        messages.success(request, 'Typ pojištění byl úspěšně odstraněn.')
+        messages.success(request, 'Typ pojištění byl úspěšně odebrán.')
         return redirect('insurance_list')
 
     return render(request, 'insurance/confirm_delete.html', {'object': insurance_type, 'type': 'typ pojištění'})
@@ -116,8 +123,10 @@ def add_or_edit_coverage(request, insurance_id, coverage_id=None):
     insurance_type = get_object_or_404(InsuranceType, id=insurance_id)
     if coverage_id:
         coverage = get_object_or_404(InsuranceCoverage, id=coverage_id)
+        is_edit = True
     else:
         coverage = InsuranceCoverage(insurance_type=insurance_type)
+        is_edit = False
 
     if request.method == 'POST':
         form = InsuranceCoverageForm(request.POST, instance=coverage)
@@ -125,13 +134,15 @@ def add_or_edit_coverage(request, insurance_id, coverage_id=None):
             coverage = form.save(commit=False)
             coverage.insurance_type = insurance_type
             coverage.save()
-            messages.success(request, 'Pojistné krytí bylo úspěšně uloženo.')
+            if is_edit:
+                messages.success(request, 'Pojistné krytí bylo úspěšně aktualizováno.')
+            else:
+                messages.success(request, 'Pojistné krytí bylo úspěšně přidáno.')
             return redirect('insurance_coverage_list', id=insurance_id)
     else:
         form = InsuranceCoverageForm(instance=coverage)
 
     return render(request, 'insurance/insurance_coverage_form.html', {'form': form, 'insurance_type': insurance_type})
-
 
 
 # Odstranění pojistného krytí
@@ -140,39 +151,68 @@ def delete_coverage(request, insurance_id, coverage_id):
 
     if request.method == 'POST':
         coverage.delete()
-        messages.success(request, 'Pojistné krytí bylo úspěšně odstraněno.')
+        messages.success(request, 'Pojistné krytí bylo úspěšně odebráno.')
         return redirect('insurance_coverage_list', id=insurance_id)
 
     return render(request, 'insurance/confirm_delete.html', {'object': coverage, 'type': 'pojistné krytí'})
 
 
 # Přidání pojistky k pojištěnci
-def add_policy(request, insured_id):
+def add_policy_from_insured(request, insured_id):
     insured_person = get_object_or_404(InsuredPerson, id=insured_id)
+
     if request.method == 'POST':
-        form = PolicyForm(request.POST)
+        form = PolicyFormFromInsured(request.POST)
         if form.is_valid():
             policy = form.save(commit=False)
             policy.insured_person = insured_person
+            policy.premium = policy.insurance_coverage.premium
             policy.save()
-            messages.success(request, 'Pojistka byla úspěšně vytvořena.')
+            messages.success(request, 'Pojištění bylo úspěšně zřízeno.')
             return redirect('insured_detail', id=insured_id)
     else:
-        form = PolicyForm()
+        form = PolicyFormFromInsured()
 
     return render(request, 'insurance/policy_form.html', {'form': form, 'insured_person': insured_person})
+
+
+def add_policy_from_coverage(request, coverage_id):
+    insurance_coverage = get_object_or_404(InsuranceCoverage, id=coverage_id)
+
+    if request.method == 'POST':
+        form = PolicyFormFromCoverage(request.POST, insurance_coverage=insurance_coverage)
+        if form.is_valid():
+            policy = form.save(commit=False)
+            policy.insurance_coverage = insurance_coverage
+            policy.premium = insurance_coverage.premium  # Automaticky nastavena již ve formuláři
+            policy.save()
+            messages.success(request, 'Pojištění bylo úspěšně zřízeno.')
+            return redirect('insured_detail', id=policy.insured_person.id)
+    else:
+        form = PolicyFormFromCoverage(insurance_coverage=insurance_coverage)
+
+    return render(request, 'insurance/policy_form.html', {'form': form, 'insurance_coverage': insurance_coverage})
 
 
 # Editace pojistky
 def edit_policy(request, id):
     policy = get_object_or_404(Policy, id=id)
+
+    # Rozhodnutí, který formulář použít
+    if policy.insured_person:
+        form_class = PolicyFormFromInsured
+    else:
+        form_class = PolicyFormFromCoverage
+
     if request.method == 'POST':
-        form = PolicyForm(request.POST, instance=policy)
+        form = form_class(request.POST, instance=policy)
         if form.is_valid():
             form.save()
+            messages.success(request, 'Pojištění bylo úspěšně aktualizováno.')
             return redirect('insured_detail', id=policy.insured_person.id)
     else:
-        form = PolicyForm(instance=policy)  # Načtení stávajících hodnot do formuláře
+        form = form_class(instance=policy)
+
     return render(request, 'insurance/policy_form.html', {'form': form, 'insured_person': policy.insured_person})
 
 
@@ -187,29 +227,17 @@ def insurance_coverage_list(request, id):
     })
 
 
-# Editace pojistky
-def add_policy(request, insured_id):
-    insured_person = get_object_or_404(InsuredPerson, id=insured_id)
-    if request.method == 'POST':
-        form = PolicyForm(request.POST)
-        if form.is_valid():
-            policy = form.save(commit=False)
-            # Nastavení pojistného (premium)
-            policy.premium = policy.insurance_coverage.premium  # Nebo jiný logický výpočet
-            policy.insured_person = insured_person
-            policy.save()
-            return redirect('insured_detail', id=insured_id)
-    else:
-        form = PolicyForm()
-    return render(request, 'insurance/policy_form.html', {'form': form, 'insured_person': insured_person})
-
-
 # Odstranění pojistky
 def delete_policy(request, id):
     policy = get_object_or_404(Policy, id=id)
     insured_id = policy.insured_person.id
-    policy.delete()
-    return redirect('insured_detail', id=insured_id)
+
+    if request.method == 'POST':
+        policy.delete()
+        messages.success(request, 'Pojištění bylo úspěšně odebráno.')
+        return redirect('insured_detail', id=insured_id)
+
+    return render(request, 'insurance/confirm_delete.html', {'object': policy})
 
 
 # Jak na pojištění
